Rename FluentS3Client to S3Client
- Delete deprecated S3Client and replace with FluentS3Client
- Rename FluentS3Client class and files to S3Client for simplicity
- Update all tests and documentation to use new S3Client name
- Maintain same fluent API functionality with cleaner naming
